rename repo from observations to hydropandas
