+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts=--durations=0 --cov-report xml:coverage.xml --cov observations
+addopts=--durations=0 --cov-report xml:coverage.xml --cov hydropandas
