+++ tests/test_002_obs_objects.py
@@ -6,8 +6,8 @@ Created on Fri Jan 31 13:26:04 2020
 """
 
 # import os
-from observations import observation as obs
-from observations import obs_collection as oc
+from hydropandas import observation as obs
+from hydropandas import obs_collection as oc
 import numpy as np
 import pandas as pd
 # import sys
