+++ tests/test_005_io_dino.py
@@ -6,8 +6,8 @@ Created on Mon Jun 24 11:43:27 2019
 """
 # import sys
 # sys.path.insert(1, "..")
-from observations.io import io_dino
-from observations import observation as obs
+from hydropandas.io import io_dino
+from hydropandas import observation as obs
 # import os
 
 # TEST_DIR = os.path.dirname(os.path.abspath(__file__))
